Move types out of GitDiff
