+++ packages/git/src/types.ts
@@ -1,8 +1,7 @@
-import * as Diff2Html from 'diff2html/lib-esm/types'
-
 export * from './GitRefs.types'
 export * from './GitCommits.types'
 export * from './Watcher.types'
+export * from './GitDiff.types'
 
 export interface StatusFile {
   path: string
@@ -13,25 +12,6 @@ export interface StatusFile {
   isDeleted: boolean
   isModified: boolean
   isRenamed: boolean
-}
-
-type Modify<T, R> = Omit<T, keyof R> & R
-export interface DiffStats {
-  insertions: number
-  deletions: number
-  filesChanged: number
-}
-export type DiffFile = Modify<
-  Diff2Html.DiffFile,
-  {
-    newName: string | null
-    oldName: string | null
-    isModified: boolean
-  }
->
-export interface DiffResult {
-  stats: DiffStats
-  files: DiffFile[]
 }
 
 export type GitFileOp =
